Refactoring error codes
